Initial commit that parses lines and builds instructions
The current code properly reads in a file and parses the lines to
populate the Lines vector with Instructions. Errors are raised for
invalid instructions. Currently, only Add and Sub instructions are
supported.

Next steps will be to actually convert instructions into opcodes
and do so for the entire instruction set.